Move sessions related functions to session.rs file
